Move tests in separate module.
