+++ exonum/src/events/noise/mod.rs
@@ -189,262 +189,3 @@ fn write(
     message.extend_from_slice(&buf[0..len]);
     Box::new(write_all(sock, message))
 }
-
-#[cfg(test)]
-mod tests {
-    use futures::sync::{mpsc, mpsc::Receiver, mpsc::Sender};
-    use futures::{done, Future, Sink, Stream};
-    use tokio_core::net::{TcpListener, TcpStream};
-    use tokio_core::reactor::Core;
-    use tokio_timer::{TimeoutStream, Timer};
-
-    use std::io::{self, Result as IoResult};
-    use std::net::SocketAddr;
-    use std::thread;
-    use std::time::Duration;
-
-    use crypto::{gen_keypair_from_seed, Seed};
-    use events::error::{into_other, log_error};
-    use events::noise::wrapper::{NoiseWrapper, NOISE_MAX_MESSAGE_LENGTH};
-    use events::noise::{Handshake, HandshakeChannel, HandshakeParams, HandshakeResult,
-                        NoiseHandshake};
-
-    #[derive(Debug, PartialEq, Copy, Clone)]
-    pub enum HandshakeStep {
-        None,
-        EphemeralKeyExchange(u8, usize),
-        StaticKeyExchange(u8, usize),
-    }
-
-    const MAX_MESSAGE_LEN: usize = 1024;
-
-    const EMPTY_MESSAGE: usize = 0;
-    const STANDARD_MESSAGE: usize = MAX_MESSAGE_LEN;
-    const LARGE_MESSAGE: usize = NOISE_MAX_MESSAGE_LENGTH + 1;
-
-    impl HandshakeParams {
-        fn default_test_params() -> Self {
-            let (public_key, secret_key) = gen_keypair_from_seed(&Seed::new([0; 32]));
-            HandshakeParams {
-                max_message_len: MAX_MESSAGE_LEN as u32,
-                public_key,
-                secret_key,
-            }
-        }
-    }
-
-    #[test]
-    fn test_noise_normal_handshake() {
-        let addr: SocketAddr = "127.0.0.1:45001".parse().unwrap();
-        let addr2 = addr.clone();
-
-        let (sender, receiver) = mpsc::channel(0);
-        let (err_sender, _) = mpsc::channel::<IoResult<()>>(0);
-        let receiver = add_timeout_millis(receiver, 500);
-        thread::spawn(move || run_handshake_listener(&addr2, sender.clone(), err_sender));
-
-        // Use first handshake only to connect.
-        let _res = send_handshake(&addr, HandshakeStep::None);
-        receiver.wait().next();
-
-        let res = send_handshake(&addr, HandshakeStep::None);
-        assert!(res.is_ok());
-    }
-
-    #[test]
-    #[should_panic(expected = "WrongMessageLength(0)")]
-    fn test_noise_handshake_errors_ee_empty() {
-        let addr: SocketAddr = "127.0.0.1:45003".parse().unwrap();
-        let step = HandshakeStep::EphemeralKeyExchange(1, EMPTY_MESSAGE);
-        let res = wait_for_handshake_result(&addr, step);
-        res.unwrap()
-    }
-
-    #[test]
-    #[should_panic(expected = "WrongMessageLength(0)")]
-    fn test_noise_handshake_errors_es_empty() {
-        let addr: SocketAddr = "127.0.0.1:45004".parse().unwrap();
-        let step = HandshakeStep::StaticKeyExchange(2, EMPTY_MESSAGE);
-        let res = wait_for_handshake_result(&addr, step);
-        res.unwrap()
-    }
-
-    #[test]
-    #[should_panic(expected = "early eof")]
-    fn test_noise_handshake_errors_ee_standard() {
-        let addr: SocketAddr = "127.0.0.1:45005".parse().unwrap();
-        let step = HandshakeStep::EphemeralKeyExchange(1, STANDARD_MESSAGE);
-        let res = wait_for_handshake_result(&addr, step);
-        res.unwrap()
-    }
-
-    #[test]
-    #[should_panic(expected = "early eof")]
-    fn test_noise_handshake_errors_es_standard() {
-        let addr: SocketAddr = "127.0.0.1:45006".parse().unwrap();
-        let step = HandshakeStep::StaticKeyExchange(2, STANDARD_MESSAGE);
-        let res = wait_for_handshake_result(&addr, step);
-        res.unwrap()
-    }
-
-    #[test]
-    #[should_panic(expected = "early eof")]
-    fn test_noise_handshake_errors_ee_large() {
-        let addr: SocketAddr = "127.0.0.1:45007".parse().unwrap();
-        let step = HandshakeStep::EphemeralKeyExchange(1, LARGE_MESSAGE);
-        let res = wait_for_handshake_result(&addr, step);
-        res.unwrap()
-    }
-
-    #[test]
-    #[should_panic(expected = "early eof")]
-    fn test_noise_handshake_errors_se_large() {
-        let addr: SocketAddr = "127.0.0.1:45008".parse().unwrap();
-        let step = HandshakeStep::StaticKeyExchange(2, LARGE_MESSAGE);
-        let res = wait_for_handshake_result(&addr, step);
-        res.unwrap()
-    }
-
-    fn wait_for_handshake_result(addr: &SocketAddr, step: HandshakeStep) -> IoResult<()> {
-        let addr2 = addr.clone();
-
-        let (sender, receiver) = mpsc::channel(0);
-        let (err_sender, err_receiver) = mpsc::channel::<IoResult<()>>(0);
-        let receiver = add_timeout_millis(receiver, 500);
-        let err_receiver = add_timeout_millis(err_receiver, 500);
-        thread::spawn(move || run_handshake_listener(&addr2, sender, err_sender));
-
-        // Use first handshake only to connect.
-        let _res = send_handshake(&addr, HandshakeStep::None);
-        receiver.wait().next();
-
-        let res = send_handshake(&addr, step);
-        assert!(res.is_err());
-
-        err_receiver.wait().next().unwrap().unwrap()
-    }
-
-    fn add_timeout_millis<T>(receiver: Receiver<T>, millis: u64) -> TimeoutStream<Receiver<T>> {
-        let timer = Timer::default();
-        timer.timeout_stream(receiver, Duration::from_millis(millis))
-    }
-
-    fn run_handshake_listener(
-        addr: &SocketAddr,
-        sender: Sender<()>,
-        err_sender: Sender<IoResult<()>>,
-    ) -> Result<(), io::Error> {
-        let mut core = Core::new().unwrap();
-        let handle = core.handle();
-        let params = HandshakeParams::default_test_params();
-
-        let fut_stream = TcpListener::bind(addr, &handle).unwrap();
-        let fut = fut_stream
-            .incoming()
-            .for_each(|(stream, _)| {
-                let sender = sender.clone();
-                let err_sender = err_sender.clone();
-                let send = sender.send(()).map(|_| ()).map_err(log_error);
-                handle.spawn(send);
-
-                let handshake = NoiseHandshake::new();
-                let handshake = handshake.listen(&params, stream);
-                let reader = handshake
-                    .and_then(|_| Ok(()))
-                    .or_else(|e| err_sender.send(Err(e)).map(|_| ()))
-                    .map_err(|e| log_error(e));
-                handle.spawn(reader);
-                Ok(())
-            })
-            .map_err(|e| into_other(e));
-
-        core.run(fut)
-    }
-
-    fn send_handshake(addr: &SocketAddr, step: HandshakeStep) -> Result<(), io::Error> {
-        let mut core = Core::new().unwrap();
-        let handle = core.handle();
-        let params = HandshakeParams::default_test_params();
-
-        let stream = TcpStream::connect(&addr, &handle)
-            .and_then(|sock| match step {
-                HandshakeStep::None => {
-                    let handshake = NoiseHandshake::new();
-                    handshake.send(&params, sock)
-                }
-                _ => {
-                    let error_handshake = NoiseErrorHandshake::new(step);
-                    error_handshake.send(&params, sock)
-                }
-            })
-            .map(|_| ())
-            .map_err(into_other);
-
-        core.run(stream)
-    }
-
-    #[derive(Debug, Copy, Clone)]
-    pub struct ErrorChannel {
-        step: HandshakeStep,
-        current_step: u8,
-    }
-
-    impl ErrorChannel {
-        fn new(step: HandshakeStep, current_step: u8) -> Self {
-            ErrorChannel { step, current_step }
-        }
-    }
-
-    impl HandshakeChannel for ErrorChannel {
-        fn read_handshake_msg(
-            &self,
-            input: &[u8],
-            noise: &mut NoiseWrapper,
-        ) -> Box<Future<Item = (usize, Vec<u8>), Error = io::Error>> {
-            let res = noise.read_handshake_msg(input);
-            Box::new(done(res.map_err(|e| e.into())))
-        }
-
-        fn write_handshake_msg(
-            &mut self,
-            noise: &mut NoiseWrapper,
-        ) -> Box<Future<Item = (usize, Vec<u8>), Error = io::Error>> {
-            let res = match &self.step {
-                // Write message filled with zeros, instead of real handshake message.
-                &HandshakeStep::EphemeralKeyExchange(cs, size)
-                | &HandshakeStep::StaticKeyExchange(cs, size) if cs == self.current_step =>
-                {
-                    println!("size {:?}", size);
-                    Ok((size, vec![0; size]))
-                }
-                _ => noise.write_handshake_msg(),
-            };
-
-            self.current_step += 1;
-            Box::new(done(res.map_err(|e| e.into())))
-        }
-    }
-
-    #[derive(Debug)]
-    pub struct NoiseErrorHandshake {
-        channel: ErrorChannel,
-    }
-
-    impl NoiseErrorHandshake {
-        fn new(step: HandshakeStep) -> Self {
-            NoiseErrorHandshake {
-                channel: ErrorChannel::new(step, 1),
-            }
-        }
-    }
-
-    impl Handshake for NoiseErrorHandshake {
-        fn listen(self, params: &HandshakeParams, stream: TcpStream) -> HandshakeResult {
-            super::listen_handshake(stream, &params, &self.channel)
-        }
-
-        fn send(self, params: &HandshakeParams, stream: TcpStream) -> HandshakeResult {
-            super::send_handshake(stream, &params, &self.channel)
-        }
-    }
-}
